refactor: extract Level II market data to market.level2 (#578)
