+++ src/open_stocks_mcp/server/app.py
@@ -34,6 +34,7 @@ from open_stocks_mcp.tools.market.earnings import (
     get_stock_events,
     get_stock_splits,
 )
+from open_stocks_mcp.tools.market.level2 import get_stock_level2_data
 from open_stocks_mcp.tools.market.movers import (
     get_stocks_by_tag,
     get_top_100,
@@ -81,9 +82,6 @@ from open_stocks_mcp.tools.robinhood_dividend_tools import (
     get_interest_payments,
     get_stock_loan_payments,
     get_total_dividends,
-)
-from open_stocks_mcp.tools.robinhood_market_data_tools import (
-    get_stock_level2_data,
 )
 from open_stocks_mcp.tools.robinhood_order_tools import (
     get_options_orders,
